moved batch extract/reduction functions to a separate file
