+++ panoptes_aggregation/scripts/__init__.py
@@ -1,6 +1,7 @@
 from .config_workflow_panoptes import config_workflow
-from .extract_panoptes_csv import extract_csv, progressbar as pbe
-from .reduce_panoptes_csv import reduce_csv, progressbar as pbr
+from .extract_panoptes_csv import extract_csv
+from .reduce_panoptes_csv import reduce_csv
+from .batch_utils import progressbar as pbe
 from .path_type import PathType
 from .gui_overrides import gui_override, pbar_override
 from .aggregation_parser import main as parser_main
